+++ knip.json
@@ -7,6 +7,7 @@
     "./lib/db/seed.ts",
     "./lib/localisation/i18n.ts",
     "./lib/localisation/navigation.ts",
-    "./lib/utils.ts"
+    "./lib/utils.ts",
+    "**/*.stories.tsx"
   ]
 }
